+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 0.1.1
 
 * Fixed `WordHelper.strip_punctuation` method
+* Added more tests
 
 ### 0.1.0